tests requiring SectorBuilder cannot be parallelized
The sector builder uses rocksdb to store its metadata, and rocksdb uses a LOCK file to coordinate operations across threads. This file is written, by default, to ~/.filecoin/sectors/staging/LOCK. If multiple tests are run in parallel which rely on this location, bad things will happen (such as an  error).